add FindMe component
